split layer prefix from types
