+++ include/spark_dsg/scene_graph_types.h
@@ -35,7 +35,6 @@
 #pragma once
 #include <cstdint>
 #include <limits>
-#include <ostream>
 #include <string>
 
 /**
@@ -55,67 +54,6 @@ struct EdgeKey {
   NodeId k2;
 };
 
-std::ostream& operator<<(std::ostream& out, const EdgeKey& key);
-
-struct LayerKey {
-  static constexpr LayerId UNKNOWN_LAYER = std::numeric_limits<LayerId>::max();
-  LayerId layer;
-  uint32_t prefix = 0;
-  bool dynamic = false;
-
-  LayerKey();
-
-  LayerKey(LayerId layer_id);
-
-  LayerKey(LayerId layer_id, uint32_t prefix);
-
-  bool isParent(const LayerKey& other) const;
-
-  bool operator==(const LayerKey& other) const;
-
-  inline bool operator!=(const LayerKey& other) const {
-    return !this->operator==(other);
-  }
-
-  inline operator bool() const { return layer != UNKNOWN_LAYER; }
-};
-
-std::ostream& operator<<(std::ostream& out, const LayerKey& key);
-
-class LayerPrefix {
- public:
-  LayerPrefix(char key);
-
-  LayerPrefix(char key, uint32_t index);
-
-  LayerPrefix(uint32_t index);
-
-  static LayerPrefix fromId(NodeId node);
-
-  inline operator uint32_t() const { return value_.value; }
-
-  std::string str(bool with_key = true) const;
-
-  bool matches(NodeId node) const;
-
-  NodeId makeId(size_t index) const;
-
-  size_t index(NodeId node_id) const;
-
-  inline char key() const { return value_.symbol.key; }
-
-  inline uint32_t index() const { return value_.symbol.index; }
-
- private:
-  union {
-    uint32_t value;
-    struct __attribute__((packed)) {
-      uint32_t index : 24;
-      char key : 8;
-    } symbol;
-  } value_;
-};
-
 /**
  * @brief Layer enum hierarchy corresponding to original DSG paper(s)
  * @note A higher layer id corresponds to parents for interlayer edges
@@ -129,7 +67,8 @@ struct DsgLayers {
   inline const static LayerId ROOMS = 4;      //< Room node layer
   inline const static LayerId BUILDINGS = 5;  //< Building node layer
   inline const static LayerId MESH_PLACES = 20;  //< Mesh (2D) Places node layer
-  inline const static LayerId UNKNOWN = LayerKey::UNKNOWN_LAYER;  //< Catchall layer ID
+  inline const static LayerId UNKNOWN =
+      std::numeric_limits<LayerId>::max();  //< Catchall layer ID
 
   static std::string LayerIdToString(LayerId id);
   static LayerId StringToLayerId(const std::string& id_str);
