+++ include/spark_dsg/spark_dsg_fwd.h
@@ -2,12 +2,16 @@
 #include <memory>
 
 namespace spark_dsg {
+
 class Mesh;
+
 struct NodeAttributes;
+
 struct EdgeAttributes;
 struct SceneGraphEdge;
 struct EdgeContainer;
 
+class SceneGraphLayer;
 class DynamicSceneGraph;
 using DynamicSceneGraphPtr = std::shared_ptr<DynamicSceneGraph>;
