+++ tests/utest_scene_graph_layer.cpp
@@ -353,12 +353,14 @@ TEST(SceneGraphLayerTests, MergeLayerCorrect) {
     EXPECT_TRUE(layer_2.insertEdge(i - 1, i));
   }
 
-  std::map<NodeId, LayerKey> node_to_layer;
-  layer_1.mergeLayer(layer_2, {}, &node_to_layer);
+  std::vector<NodeId> new_nodes;
+  layer_1.mergeLayer(layer_2, {}, &new_nodes);
 
-  EXPECT_EQ(2u, node_to_layer.size());
   EXPECT_EQ(5u, layer_1.numNodes());
   EXPECT_EQ(4u, layer_1.numEdges());
+
+  std::vector<NodeId> expected_new_nodes{3, 4};
+  EXPECT_EQ(new_nodes, expected_new_nodes);
 
   for (size_t i = 0; i < 5; i++) {
     Eigen::Vector3d result = layer_1.getPosition(i);
@@ -366,9 +368,6 @@ TEST(SceneGraphLayerTests, MergeLayerCorrect) {
     EXPECT_NEAR(0.0, result(1), 1.0e-9);
     EXPECT_NEAR(0.0, result(2), 1.0e-9);
     EXPECT_EQ(NodeStatus::NEW, layer_1.checkNode(i));
-    if (i > 2) {
-      EXPECT_EQ(LayerKey(1), node_to_layer.at(i));
-    }
   }
 }
